requirements.txt
flask
gunicorn
wtforms
email_validator
flask-login
oauthlib
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
pyopenssl
psycopg2-binary
apscheduler
redis